requirements.txt
# FlushBot Dependencies

# Core Telegram Bot
python-telegram-bot==20.7
asyncio==3.4.3

# AI & HTTP Clients
openai==1.5.0
httpx==0.25.2
aiohttp==3.9.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Cache & Redis
redis==5.0.1
hiredis==2.2.3

# Data Processing
pandas==2.1.4
pyarrow==14.0.2
numpy==1.26.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Security & Utilities
cryptography==41.0.8
bcrypt==4.1.2
hashids==1.3.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Rate Limiting & Utils
slowapi==0.1.9
schedule==1.2.0
pytz==2023.3